Refactor ESP driver
Move WiFi stuff into a separate file.
