+++ OWLOS/src/src/config.h
@@ -73,4 +73,24 @@ OWLOS распространяется в надежде, что она буде
     #endif
 #endif
 
+#define DEFAULT_WIFI_ACCESS_POINT_AVAILABLE 1
+#define DEFAULT_WIFI_ACCESS_POINT_SSID "owlnode"
+#define DEFAULT_WIFI_ACCESS_POINT_PASSWORD  "1122334455"
+#define DEFAULT_WIFI_ACCESS_POINT_IP  "192.168.4.1"
+
+#define DEFAULT_WIFI_STATION_AVAILABLE 1
+#define DEFAULT_WIFI_STATION_SSID "Palata#13"
+#define DEFAULT_WIFI_STATION_PASSWORD "qweasdzxc1234"
+
+// WiFi properties
+#define DEFAULT_WIFI_MODE WIFI_OFF
+#define DEFAULT_WIFI_STATUS WL_DISCONNECTED
+#define DEFAULT_WIFI_STATUS_TO_STR WL_DISCONNECTED
+
+#define DEFAULT_ZERO_VALUE 0x00
+#define DEFAULT_EMPTY_STR_VALUE ""
+#define DEFAULT_ID "owlnode"
+#define DEFAULT_TOPIC "world0/area1/front1/room1/"
+
+
 //#define USE_SCRIPT
